Implémentation Q1-Q5 : assignation appels agents et IA

+++ src/main/java/tn/esprit/oussemamsehliarctic10/repositories/ICalls.java
@@ -3,5 +3,5 @@ package tn.esprit.oussemamsehliarctic10.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import tn.esprit.oussemamsehliarctic10.entities.Calls;
 
-public interface ICalls extends JpaRepository<Calls, Integer> {
+public interface ICalls extends JpaRepository<Calls, Long> {
 }
